Refactor: move FollowerState to follower_state.rs
